+++ cmd/config.go
@@ -5,7 +5,6 @@ import (
 	"log"
 	"os"
 
-	"github.com/klaudiusz-czapla/my-cloud-home-go/mch/models"
 	"github.com/spf13/cobra"
 )
