+++ mch/serde/config.go
@@ -1,0 +1,8 @@
+package serde
+
+type MchConfig struct {
+	Data struct {
+		ConfigurationID string                            `json:"configurationId"`
+		ComponentMap    map[string]map[string]interface{} `json:"componentMap"`
+	} `json:"data"`
+}
