+++ mch/serde/idtokenpayload.go
@@ -1,4 +1,4 @@
-package models
+package serde
 
 type IdTokenPayload struct {
 	Iss string `json:"iss"`
