+++ mch/utils/utils.go
@@ -4,12 +4,12 @@ import (
 	"fmt"
 
 	"github.com/golang-jwt/jwt/v4"
-	"github.com/klaudiusz-czapla/my-cloud-home-go/mch/models"
+	"github.com/klaudiusz-czapla/my-cloud-home-go/mch/serde"
 	"github.com/klaudiusz-czapla/my-cloud-home-go/utils"
 	"github.com/mitchellh/mapstructure"
 )
 
-func DecodeIdToken(tokenString string) (*jwt.MapClaims, *models.IdTokenPayload, error) {
+func DecodeIdToken(tokenString string) (*jwt.MapClaims, *serde.IdTokenPayload, error) {
 	claims := jwt.MapClaims{}
 	_, _, err := new(jwt.Parser).ParseUnverified(tokenString, &claims)
 
@@ -25,7 +25,7 @@ func DecodeIdToken(tokenString string) (*jwt.MapClaims, *models.IdTokenPayload, 
 		return nil, nil, fmt.Errorf("claims inside the token are not valid")
 	}
 
-	var idTokenPayload = models.IdTokenPayload{}
+	var idTokenPayload = serde.IdTokenPayload{}
 	mapstructure.Decode(claims, &idTokenPayload)
 
 	json, _ := utils.ToJson(&idTokenPayload)
@@ -34,7 +34,7 @@ func DecodeIdToken(tokenString string) (*jwt.MapClaims, *models.IdTokenPayload, 
 	return &claims, &idTokenPayload, nil
 }
 
-func DecodeAccessToken(tokenString string) (*jwt.MapClaims, *models.AccessTokenPayload, error) {
+func DecodeAccessToken(tokenString string) (*jwt.MapClaims, *serde.AccessTokenPayload, error) {
 	claims := jwt.MapClaims{}
 	_, _, err := new(jwt.Parser).ParseUnverified(tokenString, &claims)
 
@@ -50,7 +50,7 @@ func DecodeAccessToken(tokenString string) (*jwt.MapClaims, *models.AccessTokenP
 		return nil, nil, fmt.Errorf("claims inside the token are not valid")
 	}
 
-	var accTokenPayload = models.AccessTokenPayload{}
+	var accTokenPayload = serde.AccessTokenPayload{}
 	mapstructure.Decode(claims, &accTokenPayload)
 
 	json, _ := utils.ToJson(&accTokenPayload)
